refactor: move gmail into lib
